# doc/openapi-mcp-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MessageMedia MCP Server",
    "version": "1.0.0",
    "description": "AI-native MCP server for SMS context and message delivery using MessageMedia API."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/context": {
      "post": {
        "summary": "Get SMS context for a phone number",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone_number": {
                    "type": "string"
                  }
                },
                "required": [
                  "phone_number"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Context summary and message history",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/send": {
      "post": {
        "summary": "Send an SMS via MessageMedia",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "messages": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "destination_number": {
                          "type": "string"
                        },
                        "content": {
                          "type": "string"
                        },
                        "format": {
                          "type": "string",
                          "example": "SMS"
                        },
                        "delivery_report": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "destination_number",
                        "content"
                      ]
                    }
                  }
                },
                "required": [
                  "messages"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SMS sent response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/meta": {
      "get": {
        "summary": "Get MCP capability metadata",
        "responses": {
          "200": {
            "description": "Metadata with capabilities and supported types",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/function-schema": {
      "get": {
        "summary": "Get OpenAI-compatible tool schema",
        "responses": {
          "200": {
            "description": "Function schema array",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}